+++ libpok/include/core/thread.h
@@ -38,7 +38,7 @@ typedef struct {
   bool_t user_add;
   uint8_t finish_num;
   uint8_t miss_num;
-  uint8_t total_num;
+  uint32_t total_num;
 } pok_thread_attr_t;
 
 void pok_thread_init(void);
